change pdf renderer

+++ src/main/java/utils/FileUtil.java
@@ -13,8 +13,8 @@ public class FileUtil {
     //按文件名排序（同一目录下）
     public static int sortByName(String name1,String name2){
         //将所有非数字替换成空格，便于后续处理
-        int[] tmp1=clearDelimitor(name1);
-        int[] tmp2=clearDelimitor(name2);
+        long[] tmp1=clearDelimitor(name1);
+        long[] tmp2=clearDelimitor(name2);
         int len = Math.min(tmp1.length,tmp2.length);
         for(int i=0;i<len;i++){
             if(tmp1[i] == tmp2[i]){//这一段相同则继续
@@ -28,7 +28,7 @@ public class FileUtil {
         return 0;
     }
     //清除分隔符号，（待定）将字母转换成可比较的数字，便于进一步比较
-    private static int[] clearDelimitor(String str) {
+    private static long[] clearDelimitor(String str) {
         String[] tmpArr =  str.replaceAll("[^0-9]", " ")
                 .split(" ");
         List<String> tmp = new ArrayList<>();
@@ -38,18 +38,13 @@ public class FileUtil {
             }
         }
         int n = tmp.size();
-        int[] result=new int[n];
+        long[] result=new long[n];
         for(int i=0;i<n;i++){
-            result[i] = Integer.parseInt(tmp.get(i));
+            result[i] = Long.parseLong(tmp.get(i));
         }
         return result;
     }
-    /**
-     *
-     * @param filePath
-     * @return true on successful deletion of the file, filePath is null (or) empty.
-     *         false in all other cases.
-     */
+    //移动文件至回收站
     public static boolean moveFileToTrash(File[] filesToDel) {
         FileUtils fileUtils = FileUtils.getInstance();
         if (!fileUtils.hasTrash()) {
